% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_namespace.R
\name{append_namespace}
\alias{append_namespace}
\title{Append namespace to functions in script}
\usage{
append_namespace(text = NULL, con = NULL, overwrite = FALSE)
}
\arguments{
\item{text}{character, vector that contains script, Default: NULL}

\item{con}{character, path to file or directory that contains script, Default: NULL}

\item{overwrite}{boolean, overwrite original file, Default: FALSE}
}
\value{
character
}
\description{
Autoappend namespace to functions in script by searchpath order
}
\details{
searches for functions in the loadedNamespace and then in the remaining installed.packages
}
\examples{
txt <- '#some comment
yy <- function(a=4){
  head(runif(10),a)
  # a comment
}

zz <- function(v=10,a=8){
  head(runif(v),a)
}'

append_namespace(text=txt)
}
\seealso{
\code{\link[stringi]{stri_sub}}
 \code{\link[utils]{getParseData}}
}
\author{
Jonathan Sidi
}
