% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_collapse.R
\name{code_collapse}
\alias{code_collapse}
\title{Collapsing tags}
\usage{
code_collapse(path)
}
\arguments{
\item{path}{character, path to file(s) or directory to apply tags to}
}
\value{
character
}
\description{
Append code collapsing tags for RStudio IDE to pairs
 of function and roxygen2 documentation
}
\author{
Jonathan Sidi
}
