% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yml.R
\name{create_yml}
\alias{create_yml}
\title{Create _sinewconfig.yml}
\usage{
create_yml()
}
\value{
nothing
}
\description{
Create _sinewconfig.yml file in project root directory
}
\examples{
\dontrun{
create_yml()
}
}
\author{
Jonathan Sidi
}
