% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_utils.R
\name{doc_packages}
\alias{doc_packages}
\title{doc_packages}
\usage{
doc_packages(txt)
}
\arguments{
\item{txt}{\code{(character)} vector of document text}
}
\value{
\code{(character)} vector of all packages loaded in the doc
}
\description{
A function that finds all packages in a document based on typical methods of loading namespaces in R
}
\details{
See \href{https://github.com/yonicd/sinew/issues/61}{#61} for a discussion
}
\keyword{Internal}
