% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_to_file.R
\name{examples_to_file}
\alias{examples_to_file}
\title{Convert examples blocks in roxygen2 header to script}
\usage{
examples_to_file(input, output = "./roxy_ex_to_file.R")
}
\arguments{
\item{input}{character, file or directory}

\item{output}{character, file path of output, Default: './roxy_ex_to_file.R'}
}
\value{
writes R file to disk
}
\description{
Converts and aggregates roxygen2 examples into a
single output file.
}
\details{
If output is set to NULL then output returned as invisible character object.
}
\author{
Jonathan Sidi
}
