% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_namespace.R
\name{force_pkgs}
\alias{force_pkgs}
\title{Create lists of \code{package} exports}
\usage{
force_pkgs(packages)
}
\arguments{
\item{packages}{\code{(character)} packages to include in list, order of packages determines which function will be selected first.}
}
\value{
\code{(named list)} with package names as names as all exports as a character vector
}
\description{
Useful for supplying packages to the \code{force} argument to \code{pretty_namespace}.
}
\examples{
force_pkgs(c("dplyr"))
}
