% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interOxyAddIn.R
\name{interOxyAddIn}
\alias{interOxyAddIn}
\title{Interactive add-in}
\usage{
interOxyAddIn()
}
\value{
Nothing. Inserts roxygen2 comments in a file opened in the source editor.
}
\description{
Launches an addin for interactive selection of parameters for \code{makeOxygen}
}
\details{
Add-in launches in the viewer panel of Rstudio. Select function's/dataset's name in the source editor. Choose parameters for \code{makeOxygen}. Click Insert. Select next object's name. Rinse.Repeat. Click Quit when done with the file.
}
\examples{
interOxyAddIn() # launches interactive add-in, alternatively,
Rstudio menu Addins -> 'interactiveOxygen' 
}
\seealso{
\code{View(sinew:::oxygenAddin)}
}
\author{
Anton Grishin
}
