% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_param.R
\name{ls_param}
\alias{ls_param}
\title{Return roxygen2 parameter calls from parameter dictionary}
\usage{
ls_param(obj, dictionary = "man-roxygen/Dictionary-1.R", print = TRUE)
}
\arguments{
\item{obj}{function or name of function}

\item{dictionary}{character, path_to_dictionary, Default: 'roxygen-man/Dictionary-1.R'}

\item{print}{boolean print output to console, Default: TRUE}
}
\value{
character vector
}
\description{
Return roxygen2 parameter calls from the intersection of the parameters listed in
the package dictionary and the formals of a function
}
\examples{
repo='https://raw.githubusercontent.com/yonicd/sinew/master/'
dict_loc=file.path(repo,'man-roxygen/Dictionary-1.R')
ls_param(sinew::makeOxygen,dictionary=dict_loc)
}
\concept{utility}
