% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_import.R
\name{make_desc}
\alias{make_desc}
\title{Scrape R script to create namespace calls for R documentation}
\usage{
make_desc(script_path = "R", desc_loc = getwd())
}
\arguments{
\item{script_path}{character, directory path to R scripts. Default: 'R'}

\item{desc_loc}{character, path to DESCRIPTION file, if not NULL then the Imports fields in
the DESCRIPTION file, Default: getwd()}
}
\description{
Scrape r script to create namespace calls for roxygen2, namespace or description files
}
\examples{
\dontrun{
make_desc()
}
}
