% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_namespace.R
\name{make_force_packages}
\alias{make_force_packages}
\title{Create lists of \code{package} exports}
\usage{
make_force_packages(packages)
}
\arguments{
\item{packages}{\code{(character)} packages to include in list. When duplicate function names exist the order of packages determines which function will be selected first - IE the first package with the function name will include that function, the second package with the function name will not have it listed.}
}
\value{
\code{(named list)} with package names as names as all exports as a character vector
}
\description{
Useful for supplying packages to the \code{force} argument to \code{pretty_namespace}.
}
\examples{
make_force_packages(c("utils"))
}
