% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDictionary.R
\name{makeDictionary}
\alias{makeDictionary}
\title{Parse package R files to create dictionary of unique parameter definitions}
\usage{
makeDictionary(path, save_path = FALSE)
}
\arguments{
\item{path}{character or character vector of paths to files to parse}

\item{save_path}{boolean that allows for function to write template to man-roxygen subdirectory, Default: FALSE}
}
\value{
character/character vector of intersecting parameters
}
\description{
Given list of R files function returns roxygen2 template consisting of intersecting
parameter definitions
}
\examples{
makeDictionary('./R')
}
