% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOxygen.R
\name{makeOxygen}
\alias{makeOxygen}
\title{Creates skeleton Roxygen with information from within function script}
\usage{
makeOxygen(obj, add_default = TRUE, add_fields = NULL, print = TRUE, ...)
}
\arguments{
\item{obj}{function or name of function}

\item{add_default}{logical to add defaults values to the end of the PARAM fields, Default: TRUE}

\item{add_fields}{character vector to add additional roxygen fields, Default: NULL}

\item{print}{logical write output to console, Default: TRUE}

\item{...}{arguments to be passed to makeImport}
}
\description{
Creates Roxygen skeleton including title, description, import and other fields
}
\details{
add_fields can include (concept,keyword,usage,export,details,examples). the order in add_fields
determines the order of printout.
}
\examples{
makeOxygen(stats::lm,add_default = TRUE,add_fields = c('export','examples'))
}
