% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSeeAlso.R
\name{makeSeeAlso}
\alias{makeSeeAlso}
\title{FUNCTION_TITLE}
\usage{
makeSeeAlso(obj, cutOFF = 3)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{cutOFF}{PARAM_DESCRIPTION, Default: 3}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
