% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moga.R
\name{moga}
\alias{moga}
\title{Make Oxygen Great Again}
\usage{
moga(path, ..., force.fields = NULL, dry.run = TRUE, overwrite = FALSE)
}
\arguments{
\item{path}{character path to R file}

\item{\dots}{arguments to be passed to new makeOxygen}

\item{force.fields}{character, vector a field names that are in current header that are to be updated Default: NULL}

\item{dry.run}{boolean, write lines to console the output, Default: TRUE}

\item{overwrite}{boolean, overwrite contents of input file, Default: FALSE}
}
\value{
character
}
\description{
Update/append an R file that has roxygen2 headers already with updated information
}
\details{
Cross references fields already in the roxygen2 header and adds any new ones from the updated call.
To force a change to a field add field name to force.fields.
}
\examples{

# We want to update the contents of the Roxygen2 with the new parameter "b"
# without touching the other fields

# Before
 cat(readLines(system.file('example_moga.R',package = 'sinew')),sep = '\n')

# After
 moga(system.file('example_moga.R',package = 'sinew'))
 
}
\concept{populate}
