% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_utils.R
\name{namespace_exports}
\alias{namespace_exports}
\title{namespace_exports}
\usage{
namespace_exports(ns)
}
\arguments{
\item{ns}{\code{(character)} vector of package (namespace) names}
}
\value{
\code{(list)} of character vectors containing all of a package's exported functions
}
\description{
given a named character vector of packages outputs a list of all package exports
}
\keyword{Internal}
