% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImport.R
\name{order_desc}
\alias{order_desc}
\title{Order Package Description File Fields}
\usage{
order_desc(
  path = ".",
  order_fields = c("Imports", "Suggests", "Remotes"),
  overwrite = TRUE
)
}
\arguments{
\item{path}{\code{chr} path to directory in which \emph{DESCRIPTION} file resides}

\item{order_fields}{\code{chr} Fields to be ordered alphabetically}

\item{overwrite}{\code{lgl} Whether to overwrite with the ordered values}
}
\value{
\code{msg} prints the fields to console. If \code{overwrite = TRUE} also writes the ordered fields to \emph{DESCRIPTION}
}
\description{
Order Package Description File Fields
}
