% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsFromTemplate.R
\name{paramsFromTemplate}
\alias{paramsFromTemplate}
\title{Create roxygen parameter calls from subset of template}
\usage{
paramsFromTemplate(obj, template = "man-roxygen/template.R")
}
\arguments{
\item{obj}{function or name of a function}

\item{template}{character, path_to_template, Default: 'roxygen-man/template.R'}
}
\value{
character vector
}
\description{
Create roxygen parameter calls from the intersection of the parameters listed in a template file
and the formals of a function
}
\examples{
paramsFromTemplate(sinew::makeOxygen,
template='https://raw.githubusercontent.com/yonicd/sinew/master/man-roxygen/template.R')
}
