% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_utils.R
\name{parse_check}
\alias{parse_check}
\title{parse_check}
\usage{
parse_check(p, txt, ask)
}
\arguments{
\item{p}{result of \code{pretty_parse} > \code{parse}}

\item{txt}{input text to \code{pretty_parse}}

\item{ask}{boolean, If TRUE then a \code{\link[utils]{menu}} will be created for the use to
choose between competing namespaces for a function, Default: TRUE}
}
\description{
check for fail of pretty_parse > parse, and offers to open file to offending line
}
\keyword{Internal}
