% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_addin.R
\name{pretty_addin}
\alias{pretty_addin}
\title{Interactively run pretty functions}
\usage{
pretty_addin()
}
\description{
Addin that scans the file source contents and attaches namespace
information.
}
\details{
Either saved or untitled R and Rmd files in the source editor may be used.

In R files you can highlight specific text, or not highlight at all and
the whole document will be used.

In Rmd files if you highlight text it must be within a single chunk, or not highlight at all and
the whole document will be used to create a new chunk at the top of the document with 
the libraries needed to run the document.
}
