% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_utils.R
\name{pretty_merge}
\alias{pretty_merge}
\title{pretty_merge}
\usage{
pretty_merge(e1, e2)
}
\arguments{
\item{e1}{\code{(data.frame)} typically \code{sym.funs} with list of all parsed functions in \code{txt}}

\item{e2}{\code{(list)} typically \code{force} or \code{ignore} with list of namespaces and the respective functions to force or ignore}
}
\description{
handles \code{force} and \code{ignore} arguments
}
