% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_rmd.R
\name{pretty_rmd}
\alias{pretty_rmd}
\title{Attach namespacing to Rmarkdown chunks}
\usage{
pretty_rmd(input, output = tempfile(fileext = ".Rmd"), open_output = TRUE,
  ...)
}
\arguments{
\item{input}{character, path to input Rmd file}

\item{output}{character, path to output Rmd file, Default: NULL}

\item{open_output}{boolean, open the output on.exit, Default: TRUE}

\item{\dots}{arguments to pass to pretty_namespace}
}
\value{
character
}
\description{
Apply pretty_namespace to Rmarkdown document
}
\details{
if output is NULL then the returned lines are printed to console.
}
\seealso{
\code{\link[sinew]{pretty_namespace}}
}
\author{
Jonathan Sidi
}
