% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_sinew.R
\name{pretty_sinew}
\alias{pretty_sinew}
\title{Convert File to R directory with pretty and oxygen}
\usage{
pretty_sinew(con = NULL, text = NULL, dir.out = NULL, keep.body = TRUE)
}
\arguments{
\item{con}{character, path to file or directory that contains script, Default: NULL}

\item{text}{character, vector that contains script, Default: NULL}

\item{dir.out}{character, path to save new R files, Default: NULL}

\item{keep.body}{boolean, if TRUE all non-functions will be saved to body.R in the parent
directory of dir.out, Default: TRUE}
}
\value{
Nothing, side effects is to create files
}
\description{
One function to run \link[=pretty_namespace]{pretty_namespace}, \link[=untangle]{untangle} and \link[=makeOxyFile]{makeOxyFile}
}
\details{
If dir.out is set to NULL all outputs are redirected into file.path(tempdir(),'sinew')
}
\author{
Jonathan Sidi
}
\concept{namespace}
