% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmOxygen.R
\name{rmOxygen}
\alias{rmOxygen}
\title{Remove roxygen2 Comments From an .R File}
\usage{
rmOxygen(.file)
}
\arguments{
\item{.file, }{path to an .R file, character vector of length 1}
}
\value{
Nothing. Overwrites files as a side effect
}
\description{
Strips .R files of roxygen2 style comments (#')
}
\examples{
\dontrun{
rmOxygen("./myRfunctions/function1.R")
}
}
\author{
Anton Grishin
}
