% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDefaultAddfields.R
\name{setDefaultAddfields}
\alias{setDefaultAddfields}
\title{set/change the default `add_fields` used by makeOxygen}
\usage{
setDefaultAddfields(add_fields = NULL)
}
\arguments{
\item{add_fields}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
set default values for `add_field` which are retained among different
calls to `makeOxygen`, by creating and setting a "sinew_defaultaddfields" entry in `options`
If called without arguments (the default), `add_fields` is set/reset to the default value
(corresponding to "details", "examples", "seealso", "rdname", "export")).
If an `add_field` argument is passed, the default values are changed (see examples)
}
\examples{
\dontrun{
if(interactive()){
 # set/reset add_fields to default
 setDefaultAddfields()
 makeOxygen(stats::lm)

 # change default `add_fields`: After this, all calls to `makeOxygen`
 # without an `add_fields` argument will use this setting

 setDefaultAddfields(c("details", "rdname", "export", "author"))
 makeOxygen(stats::lm)
 }
}
}
