% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_file.R
\name{split_file}
\alias{split_file}
\title{Split an R script by functions}
\usage{
split_file(text = NULL, file = "", dir.out = NULL)
}
\arguments{
\item{text}{character, vector of R commands, Default: NULL}

\item{file}{character, path to R file, Default: ''}

\item{dir.out}{character, path to save new R files, Default: NULL}
}
\value{
list of seperate functions
}
\description{
Split a R script with multiple functions into multiple single function R files.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Jonathan Sidi
}
