% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_collapse.R
\name{strip_collapse}
\alias{strip_collapse}
\title{Remove sinew collapsing tags}
\usage{
strip_collapse(path)
}
\arguments{
\item{path}{character, path to file(s) or directory to apply tags to}
}
\value{
character
}
\description{
Remove collapsing tags from a lines of code
}
\author{
Jonathan Sidi
}
