% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestTemplate.R
\name{suggestTemplate}
\alias{suggestTemplate}
\title{Parse package R files to suggest Roxygen template of intersecting parameters}
\usage{
suggestTemplate(path, cutoff = 2, save_path = FALSE)
}
\arguments{
\item{path}{character or character vector of paths to files to search}

\item{cutoff}{numeric, minimal number of intersecting parameters to return, Default: 2}

\item{save_path}{boolean that allows for function to write template to man-roxygen subdirectory, Default: FALSE}
}
\value{
character/character vector of intersecting parameters
}
\description{
Given list of R files function returns Roxygen template consisting of intersecting
parameter definitions
}
\examples{
suggestTemplate('R',cutoff = 3)
}
