% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\title{Tabular for roxygen2}
\source{
\href{https://CRAN.R-project.org/package=roxygen2/vignettes/roxygen2.html}{roxygen2 formatting}
}
\usage{
tabular(df, header = TRUE, ...)
}
\arguments{
\item{df}{data.frame to convert to table}

\item{header}{boolean to control if header is created from names(df), Default: TRUE}

\item{\dots}{arguments to pass to format}
}
\value{
character
}
\description{
Convert data.frame to roxygen2 tabular format
}
\examples{
tabular(mtcars[1:5, 1:5])
tabular(mtcars[1:5, 1:5],header=FALSE)
}
\seealso{
\link[base:format]{format}
}
\concept{utility}
