% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untangle_examples.R
\name{untangle_examples}
\alias{untangle_examples}
\title{Convert examples blocks in roxygen2 header to script}
\usage{
untangle_examples(input, output = "./roxy_ex_to_file.R")
}
\arguments{
\item{input}{character, file or directory}

\item{output}{character, file path of output, Default: './roxy_ex_to_file.R'}
}
\value{
writes R file to disk
}
\description{
Converts and aggregates roxygen2 examples into a
single output file.
}
\details{
If output is set to NULL then output returned as invisible character object.
}
\author{
Jonathan Sidi
}
