% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_namespace.R
\name{append_namespace}
\alias{append_namespace}
\title{Append namespace to functions in script}
\usage{
append_namespace(text = NULL, con = NULL, overwrite = FALSE)
}
\arguments{
\item{overwrite}{boolean, overwrite original file, Default: FALSE}

\item{file}{character, path to file or directory that contains script, Default: NULL}
}
\value{
character
}
\description{
Autoappend namespace to functions in script by searchpath order
}
\details{
searches for functions in the loadedNamespace and then in the remaining installed.packages
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_sub}}
 \code{\link[utils]{getParseData}}
}
\author{
Jonathan Sidi
}
