% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImport.R
\name{makeImport}
\alias{makeImport}
\title{Scrape R script to create namespace calls for R documentation}
\usage{
makeImport(script, cut = NULL, print = TRUE, format = "oxygen")
}
\arguments{
\item{script}{character connection to pass to readLines, can be file path, directory path, url path}

\item{cut}{integer number of functions to write as importFrom until switches to import, Default: NULL}

\item{print}{boolean print output to console, Default: TRUE}

\item{format}{character the output format must be in c('oxygen','namespace','description'), Default: 'oxygen'}
}
\description{
Scrape r script to create namespace calls for roxygen2, namespace or description files
}
\examples{
makeImport(list.files('R',full.names = TRUE),format = 'oxygen')
makeImport(list.files('R',full.names = TRUE),format = 'namespace')
makeImport(list.files('R',full.names = TRUE),format = 'description')
}
