% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImport.R
\name{makeImport}
\alias{makeImport}
\title{Scrape R script to create namespace calls for R documentation}
\usage{
makeImport(script, cut = NULL, print = TRUE, format = "oxygen",
  desc_loc = NULL)
}
\arguments{
\item{script}{character, connection to pass to readLines, can be file path, directory path, url path}

\item{cut}{integer, number of functions to write as importFrom until switches to import, Default: NULL}

\item{print}{boolean, print output to console, Default: TRUE}

\item{format}{character, the output format must be in c('oxygen','description'), Default: 'oxygen'}

\item{desc_loc}{character, path to DESCRIPTION file, if not NULL then the Imports fields in
the DESCRIPTION file, Default: NULL}
}
\description{
Scrape r script to create namespace calls for roxygen2, namespace or description files
}
\examples{
makeImport('R',format = 'oxygen')
makeImport('R',format = 'description')
}
