% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_addin.R
\name{pretty_addin}
\alias{pretty_addin}
\title{Interactively run pretty functions in R files}
\usage{
pretty_addin()
}
\description{
Addin that scans the file source contents and attaches namespace
information.
}
\details{
Either saved or untitled R files in the source editor may be used.

Highlight specific text, or not highlight at all and the whole document will be used.
}
