% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDefaultAuthor.R
\name{setDefaultAuthor}
\alias{setDefaultAuthor}
\title{set a default author for makeOxygen}
\usage{
setDefaultAuthor(author = NULL)
}
\arguments{
\item{author}{`character` or object of class `person` to be used to automatically fill
the @author field when calling `makeOxygen`. If NULL, resets the @author field to
("AUTHOR [AUTHOR_2]"), Default: NULL}
}
\description{
function allowing to set the options()$sinew_DefaultAuthor to desired value (or
reset it to default).
}
\examples{
\dontrun{
 # Create an "author" variable
 author <- c(person("John", "Smith, phD (2017)", role = "aut"),
             person(c("Mary", "D."), "Ripley", role = c("ctb")))

 # set it as default for makeOxygen
 setDefaultAuthor(author)

 # include @author as a default field
 setDefaultAddfields(c("details", "export", "author"))

 # create Rd skeleton, including automatically filled-in author
 makeOxygen(stats::lm)
 }
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
